address review: clarify op descriptions, move brush tests to own file
